Update image_analysis.rs
IT NOW BUILDS!!!

I mean, there's 16 warnings, and the process to make it build in Windows is akin to being stuck in a featureless room with sirens blaring, but it does at least build now! No, I haven't checked to see if it actually processes an image to midi lol.